Allow dynamic properties
In PHP 8.2+ deprecations are emitted due to the extensive use of dynamic properties within the library.

The "best" solution is to define properties where possible and/or use magic getters and setters. From what I can see, both require extensive refactoring of the library to achieve.

The "second best" option or last resort, and by far the easiest fix is at this stage is to simply allow dynamic properties.